Add Contact Me and Footer
